Separate test-utility into compile command and run command.
